Move classifier functions from utils to result table
